+++ calour/util.py
@@ -43,43 +43,6 @@ import scipy
 logger = getLogger(__name__)
 
 
-def _iter_fasta(fp):
-    '''Iterate over fasta file.
-
-    Fasta file must contain header line (starting with ">") and one or more sequence lines.
-
-    Parameters
-    ----------
-    fp: str
-        name of the fasta file
-
-    Yields
-    ------
-    header: str
-        the header line (without ">")
-    sequence: str
-        the sequence ('ACGT'). Both header and sequence are whitespace stripped.
-    '''
-    # skip non-header lines at beginning of file
-    with open(fp, 'r') as fl:
-        for cline in fl:
-            if cline[0] == ">":
-                title = cline[1:].rstrip()
-                break
-            logger.warning('Fasta file %s has no headers' % fp)
-            return
-
-        lines = []
-        for cline in fl:
-            if cline[0] == ">":
-                yield title, ''.join(lines)
-                lines = []
-                title = cline[1:].strip()
-                continue
-            lines.append(cline.strip())
-        yield title, "".join(lines)
-
-
 def compute_prevalence(abundance):
     '''Return the prevalence at each abundance cutoffs.
